Moving former MetaFixer class (now FsRegionsMetaRecoverer) from 'hbck2.meta' package to project's top level package.
